Organise Duke files in packages
