--- src/main/java/Task.java
@@ -1,28 +1,0 @@
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description, boolean isDone) throws DukeException {
-
-        this.description = description;
-        this.isDone = isDone;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "X" : " "); // mark done task with X
-    }
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    public void unmark() {
-        this.isDone = false;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[%s] %s", this.isDone ? "X" : " ", this.description);
-    }
-
-}
