+++ src/main/java/duke/Command.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum Command {
     EVENT,
     DEADLINE,
