+++ src/main/java/duke/Storage.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
