+++ src/main/java/duke/TaskList.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
